在 users table 新增 autoincrement 的 id 修改 reverse migration 的一些錯誤

+++ database/migrations/2017_07_17_162648_add_job_title_column_to_users_table.php
@@ -26,8 +26,10 @@ class AddJobTitleColumnToUsersTable extends Migration
      */
     public function down()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('job_title');
-        });
+        if (Schema::hasColumn('users', 'job_title')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('job_title');
+            });
+        }
     }
 }
